Change namespace to extrema
