--- top-k-bottom-k/src/main/java/org/wso2/siddhi/extension/custom/util/ListNode.java
@@ -1,54 +1,0 @@
-/*
- * Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- *
- */
-
-package org.wso2.siddhi.extension.custom.util;
-
-public class ListNode<T> {
-    private T value;
-    private ListNode<T> nextNode;
-    private ListNode<T> previousNode;
-
-    public ListNode(T value) {
-        this.value = value;
-    }
-
-    public T getValue() {
-        return value;
-    }
-
-    public void setValue(T value) {
-        this.value = value;
-    }
-
-    public ListNode<T> getNextNode() {
-        return nextNode;
-    }
-
-    public void setNextNode(ListNode<T> nextNode) {
-        this.nextNode = nextNode;
-    }
-
-    public ListNode<T> getPreviousNode() {
-        return previousNode;
-    }
-
-    public void setPreviousNode(ListNode<T> previousNode) {
-        this.previousNode = previousNode;
-    }
-}
