--- top-k-bottom-k/src/main/resources/custom.siddhiext
@@ -1,25 +1,0 @@
-#
-# Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-#
-# WSO2 Inc. licenses this file to you under the Apache License,
-# Version 2.0 (the "License"); you may not use this file except
-# in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied. See the License for the
-# specific language governing permissions and limitations
-# under the License.
-#
-
-topK=org.wso2.siddhi.extension.custom.TopKStreamProcessorExtension
-bottomK=org.wso2.siddhi.extension.custom.BottomKStreamProcessorExtension
-
-topKLengthBatch=org.wso2.siddhi.extension.custom.TopKLengthBatchStreamProcessorExtension
-topKTimeBatch=org.wso2.siddhi.extension.custom.TopKTimeBatchStreamProcessorExtension
-bottomKLengthBatch=org.wso2.siddhi.extension.custom.BottomKLengthBatchStreamProcessorExtension
-bottomKTimeBatch=org.wso2.siddhi.extension.custom.BottomKTimeBatchStreamProcessorExtension
